+++ packages/tswebextension/src/lib/mv3/background/scripting-api.ts
@@ -1,3 +1,6 @@
+import { type ScriptletData } from '@adguard/tsurlfilter';
+import { type IConfiguration } from '@adguard/scriptlets';
+
 import { appContext } from './app-context';
 import { BACKGROUND_TAB_ID } from '../../common/constants';
 
@@ -28,12 +31,77 @@ export type InsertCSSParams = {
 };
 
 /**
+ * Whether to use a blob to inject the script.
+ */
+const USE_BLOB_DEFAULT = false;
+
+/**
  * Parameters for executing script.
  */
 export type ExecuteScriptParams = {
+    /**
+     * The ID of the tab.
+     */
     tabId: number,
+
+    /**
+     * The ID of the frame.
+     */
     frameId: number,
+
+    /**
+     * The script content to be executed.
+     */
     scriptText: string,
+
+    /**
+     * Whether to use a blob to inject the script.
+     * By default, it is set to {@link USE_BLOB_DEFAULT}.
+     */
+    useBlob?: boolean,
+};
+
+/**
+ * Parameters for executing scriptlet.
+ */
+export type ExecuteScriptletParams = {
+    /**
+     * The ID of the tab.
+     */
+    tabId: number,
+
+    /**
+     * The ID of the frame.
+     */
+    frameId: number,
+
+    /**
+     * The scriptlet data to be executed.
+     */
+    scriptletData: ScriptletData,
+
+    /**
+     * The domain name of the frame.
+     */
+    domainName: string | null,
+};
+
+/**
+ * Parameters for injecting the script.
+ */
+type InjectFuncOptions = {
+    /**
+     * Unique identifier for the script.
+     * It is used to prevent multiple executions of the same script on the page.
+     */
+    uniqueIdentifier: string,
+
+    /**
+     * Whether to use a blob to inject the script.
+     * Used in rare cases when inline scripts are not allowed by the CSP policy.
+     * For example, on Facebook.
+     */
+    useBlob: boolean,
 };
 
 /**
@@ -62,15 +130,60 @@ export class ScriptingApi {
     }
 
     /**
+     * Executes a scriptlet within the scope of the page.
+     *
+     * @param params Parameters for executing the scriptlet.
+     * @param params.tabId The ID of the tab.
+     * @param params.frameId The ID of the frame.
+     * @param params.scriptletData The scriptlet data to be executed.
+     * @param params.domainName The domain name of the frame. Used for debugging.
+     * @returns Promise that resolves when the script is executed.
+     */
+    public static async executeScriptlet(
+        {
+            tabId,
+            frameId,
+            scriptletData,
+            domainName,
+        }: ExecuteScriptletParams,
+    ): Promise<void> {
+        // There is no reason to inject a script into the background page
+        if (tabId === BACKGROUND_TAB_ID) {
+            return;
+        }
+
+        const params: IConfiguration = {
+            ...scriptletData.params,
+            uniqueId: String(appContext.startTimeMs),
+            verbose: appContext.configuration?.settings.debugScriptlets || false,
+            domainName: domainName ?? undefined,
+        };
+
+        await chrome.scripting.executeScript({
+            target: { tabId, frameIds: [frameId] },
+            func: scriptletData.func,
+            injectImmediately: true,
+            world: 'MAIN',
+            args: [params, scriptletData.params.args],
+        });
+    }
+
+    /**
      * Executes a script within the scope of the page.
      *
      * @param params Parameters for executing the script.
      * @param params.tabId The ID of the tab.
      * @param params.frameId The ID of the frame.
      * @param params.scriptText The script content to be executed.
+     * @param params.useBlob Whether to use a blob to inject the script.
      * @returns Promise that resolves when the script is executed.
      */
-    public static async executeScript({ tabId, frameId, scriptText }: ExecuteScriptParams): Promise<void> {
+    public static async executeScript({
+        tabId,
+        frameId,
+        scriptText,
+        useBlob = USE_BLOB_DEFAULT,
+    }: ExecuteScriptParams): Promise<void> {
         // There is no reason to inject a script into the background page
         if (tabId === BACKGROUND_TAB_ID) {
             return;
@@ -81,13 +194,18 @@ export class ScriptingApi {
          * To prevent multiple executions, this function checks if the script has already been executed.
          *
          * @param scriptText The script content to be executed.
-         * @param executedFlag A flag to check if the script has already been executed.
+         * @param options Options for the inject function {@link InjectFuncOptions}.
          */
-        // eslint-disable-next-line @typescript-eslint/no-shadow
-        function injectFunc(scriptText: string, executedFlag: string): void {
-            // We don't care about types here
+        function injectFunc(
+            // eslint-disable-next-line @typescript-eslint/no-shadow
+            scriptText: string,
+            options: InjectFuncOptions,
+        ): void {
+            const flag = 'done';
+
+            // Do not care about types here
             // @ts-ignore
-            if (window[executedFlag]) {
+            if (Window.prototype.toString[options.uniqueIdentifier] === flag) {
                 return;
             }
 
@@ -109,24 +227,19 @@ export class ScriptingApi {
                 policy = trustedTypes.createPolicy(AG_POLICY_NAME, AGPolicy);
             }
 
+            /**
+             * Injects the script into the page using a blob.
+             * This method is used when inline scripts are not allowed by the CSP policy.
+             * We use it only in rare cases since it is not so fast as the injecting via script tag with textContent.
+             * @param scriptText The script content to be executed.
+             */
             // eslint-disable-next-line @typescript-eslint/no-shadow
-            const injectViaScriptTag = (scriptText: string): void => {
+            const injectViaScriptTagWithBlob = (scriptText: string): void => {
                 const scriptTag = document.createElement('script');
 
-                let blob;
-                let url;
-                try {
-                    // This method is used to inject the script if inline scripts are not allowed by the CSP policy.
-                    // It is attempted first because it will throw an error if it fails.
-                    blob = new Blob([scriptText], { type: 'text/javascript; charset=utf-8' });
-                    url = URL.createObjectURL(blob);
-                    scriptTag.src = policy.createScriptURL(url);
-                } catch (e) {
-                    // This method does not throw an error in case of CSP policy restrictions,
-                    // which is why it is used in the catch block.
-                    scriptTag.setAttribute('type', 'text/javascript');
-                    scriptTag.appendChild(document.createTextNode(scriptText));
-                }
+                const blob = new Blob([scriptText], { type: 'text/javascript; charset=utf-8' });
+                const url = URL.createObjectURL(blob);
+                scriptTag.src = policy.createScriptURL(url);
 
                 const parent = document.head || document.documentElement;
                 if (parent) {
@@ -140,6 +253,27 @@ export class ScriptingApi {
                 if (scriptTag.parentNode) {
                     scriptTag.parentNode.removeChild(scriptTag);
                 }
+            };
+
+            /**
+             * Injects the script into the page using a script tag.
+             * @param scriptText The script content to be executed.
+             */
+            // eslint-disable-next-line @typescript-eslint/no-shadow
+            const injectViaScriptTag = (scriptText: string): void => {
+                const scriptTag = document.createElement('script');
+
+                scriptTag.setAttribute('type', 'text/javascript');
+                scriptTag.appendChild(document.createTextNode(scriptText));
+
+                const parent = document.head || document.documentElement;
+                if (parent) {
+                    parent.appendChild(scriptTag);
+                }
+
+                if (scriptTag.parentNode) {
+                    scriptTag.parentNode.removeChild(scriptTag);
+                }
 
                 if (scriptTag) {
                     scriptTag.textContent = policy.createScript('');
@@ -147,27 +281,31 @@ export class ScriptingApi {
             };
 
             const scriptTextWithPolicy = policy.createScript(scriptText);
-            try {
-                // eslint-disable-next-line no-eval
-                eval(scriptTextWithPolicy);
-            } catch (e) {
+            if (options.useBlob) {
+                injectViaScriptTagWithBlob(scriptTextWithPolicy);
+            } else {
                 injectViaScriptTag(scriptTextWithPolicy);
             }
 
-            // We don't care about types here
-            // @ts-ignore
-            window[executedFlag] = true;
+            Object.defineProperty(Window.prototype.toString, options.uniqueIdentifier, {
+                value: flag,
+                enumerable: false,
+                writable: false,
+                configurable: false,
+            });
         }
 
-        // TODO figure out how to avoid double injection without polluting global scope
-        const executedFlag = `executed${appContext.startTimeMs}`;
+        const uniqueIdentifier = `AG_done_${appContext.startTimeMs}`;
 
         await chrome.scripting.executeScript({
             target: { tabId, frameIds: [frameId] },
             func: injectFunc,
             injectImmediately: true,
             world: 'MAIN', // ISOLATED doesn't allow to execute code inline
-            args: [scriptText, executedFlag],
+            args: [scriptText, {
+                uniqueIdentifier,
+                useBlob,
+            }],
         });
     }
 }